requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.121.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
-e git+https://github.com/AmirmuhammadShk/Jobuine.git@058eab29a569bb8ed7fda0ff0371137d1b7c0e54#egg=jobuine
MarkupSafe==3.0.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
